Changes considering remarks. Added Spring property source.

+++ service/src/main/java/com/epam/esm/services/service/tag/TagService.java
@@ -1,42 +1,40 @@
-package com.epam.esm.services.service;
+package com.epam.esm.services.service.tag;
 
 import com.epam.esm.persistence.util.AscDesc;
-import com.epam.esm.persistence.dao.EntityDAO;
 import com.epam.esm.persistence.util.EntityFinder;
-import com.epam.esm.persistence.dao.TagDAO;
+import com.epam.esm.persistence.dao.tag.TagDAO;
 import com.epam.esm.model.entity.Certificate;
 import com.epam.esm.model.entity.Tag;
 import com.epam.esm.persistence.util.TagFinder;
 import com.epam.esm.persistence.exceptions.DAOException;
 import com.epam.esm.services.exceptions.ServiceException;
 import com.epam.esm.services.exceptions.ValidationException;
+import com.epam.esm.services.service.EntityService;
 import com.epam.esm.services.validator.EntityValidator;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
 import java.util.Map;
+import java.util.Optional;
 
 @Service
 public class TagService implements EntityService<Tag> {
 
-    private static final String WRONG_DAO_MESSAGE = "Wrong DAO";
-    private static TagService INSTANCE = new TagService();
-    private EntityDAO<Tag> dao;
+    @Autowired
+    private static TagService INSTANCE;
+
+    private TagDAO dao;
     private EntityValidator<Tag> validator;
+    private TagFinder finder;
 
     @Autowired
-    public void setDao(EntityDAO<Tag> dao) {
+    private TagService(TagDAO dao,
+                       EntityValidator<Tag> validator,
+                       TagFinder finder) {
         this.dao = dao;
-    }
-
-    @Autowired
-    public void setValidator(EntityValidator<Tag> validator) {
         this.validator = validator;
-    }
-
-    private TagService() {
+        this.finder = finder;
     }
 
     public static EntityService<Tag> getInstance() {
@@ -65,6 +63,9 @@ public class TagService implements EntityService<Tag> {
     @Override
     public void delete(int id) throws ServiceException {
         try {
+            if (dao.read(id) == null) {
+                throw new ServiceException("Entity does not exist");
+            }
             dao.delete(id);
         } catch (DAOException e) {
             throw new ServiceException(e);
@@ -92,47 +93,37 @@ public class TagService implements EntityService<Tag> {
 
     public Collection<Tag> findAll(Certificate certificate) throws ServiceException {
         try {
-            if (dao instanceof TagDAO) {
-                return ((TagDAO)dao).findAllByCertificate(certificate);
-            } else {
-                throw new ServiceException(WRONG_DAO_MESSAGE);
-            }
+            return dao.findAllByCertificate(certificate);
         } catch (DAOException e) {
             throw new ServiceException(e);
         }
     }
 
     public Collection<Tag> findBy(EntityFinder<Tag> entityFinder) throws ServiceException {
-        if (dao instanceof TagDAO) {
-            try {
-                return ((TagDAO)dao).findBy(entityFinder);
-            } catch (DAOException e) {
-                throw new ServiceException(e);
-            }
-        } else {
-            throw new ServiceException(WRONG_DAO_MESSAGE);
+        try {
+            return ((TagDAO) dao).findBy(entityFinder);
+        } catch (DAOException e) {
+            throw new ServiceException(e);
         }
     }
 
-    public Collection<Tag> find(String name, Map<String, String> sorting) throws ServiceException {
-        if (dao instanceof TagDAO) {
-            try {
-                TagFinder finder = new TagFinder();
-                if (StringUtils.isNotEmpty(name)) {
-                    finder.findByName(name);
-                }
-                if (sorting != null) {
-                    if (sorting.containsKey(sorting.get("sort-by-name"))) {
-                        finder.sortByName(AscDesc.getValue(sorting.get("sort-by-name")));
+    public Collection<Tag> find(Map<String, String> params) throws ServiceException {
+        finder.newFinder();
+        for (String key : params.keySet()) {
+            if (key.contains("sort")) {
+                finder.sortBy(TagSortingParameters.getEntryByParameter(key).getValue(),
+                        AscDesc.valueOf(params.get(key)));
+            } else {
+                Optional<TagSearchParameters> optional =
+                        Optional.ofNullable(TagSearchParameters.getEntryByParameter(key));
+                if (optional.isPresent()) {
+                    if (optional.get() == TagSearchParameters.NAME) {
+                        finder.findByName(params.get(key));
                     }
                 }
-                return ((TagDAO)dao).findBy(finder);
-            } catch (DAOException e) {
-                throw new ServiceException(e);
             }
-        } else {
-            throw new ServiceException(WRONG_DAO_MESSAGE);
         }
+        return findBy(finder);
     }
 
 }
